Refactor: make ActiveGameCard collapsible and rename ContinueGameCard
This commit refactors the `ContinueGameCard` into a new `ActiveGameCard` composable that can be expanded or collapsed by the user.

**Key Changes:**

- **`ContinueGameCard.kt` -> `ActiveGameCard.kt`:**
    - Renamed the file and the main composable.
    - Introduced `isExpanded` state to control the card's appearance.
    - Added `onToggle` lambda to handle expand/collapse actions.
    - Implemented `CollapsedContent` and `ExpandedContent` private composables to display different UIs based on the `isExpanded` state.
    - **`CollapsedContent`:**
        - Shows "Active Game" title.
        - Includes a `FilledIconButton` with a play icon (or trophy icon if completed) to continue/view the game.
        - Includes an `IconButton` with a "KeyboardArrowDown" icon to expand the card.
    - **`ExpandedContent`:**
        - Similar to the original `ContinueGameCard` layout.
        - Shows "Active Game" title, game details (Difficulty, Size, Time).
        - Includes an `IconButton` with a pause icon (or trophy icon if completed) next to the title.
        - Includes an `IconButton` with a "KeyboardArrowUp" icon to collapse the card.
        - Includes a full-width "Continue" / "View Board" button.
    - Uses `animateContentSize` for smooth transitions between collapsed and expanded states.
- **`SudokuCreatorScreen.kt`:**
    - Replaced `ContinueGameCard` with `ActiveGameCard`.
    - Added `rememberSaveable` state for `isExpanded` to preserve its state across recompositions and configuration changes.
    - Passed the `isExpanded` state and a toggle lambda to the `ActiveGameCard`.
- **`feature/creator/src/main/res/drawable/play_arrow.xml`:**
    - Added a new vector drawable for a play arrow icon, used in the `CollapsedContent` of the `ActiveGameCard`.
- **Previews:**
    - Added new previews (`CollapsibleActiveGameCardPreviewExpanded` and `CollapsibleActiveGameCardPreviewCollapsed`) to showcase both states of the `ActiveGameCard`.
